Refactor parameter VMs for reuse
